Added fields to Species Model for fn012 and procVal constraints
This commit adds fields the species table that capture the extents of
length, weight and condition factors for each species. It also
includes fields that can be used to estimate total length from fork
length or vice versa. A new api endpoint was added for these flen2tlen
values and filters were added for both species list and flen2tlen list
endpoints.  Several unit tests are also include to ensure that the
filters work as expected and that the returned objects have the
expected attributes. All tests continue to pass.