Added tables and constrain for ManUnitType and LakeManUnitType
This commit adds two new tables one for management unit types, and one
for Lake-Management Unit types, the later also includes a partial
constraint to ensure that there only ever one management unit type that
is identified as primary within a lake.  Unit tests were added to ensure
that the string representation is accurate, slugs are created on save
and that the partial constraint works as expected.  All previous
tests continue to pass.