Added tests and factories for lookup tables. Remove last migration
This commit adds factories for the cover and bottom type models.  It
also includes tests to verify that all of the properties and methods
of the lookup table abstract model work as expected.  The last
migration was removed from the repository - it populated the bottom
and cover tables - but these values were populated in tests too which
resulted in unexpected behaviour.  All tests continue to pass.