Added Taxon model, views, templates and api using Django treebeard
This commit integrates all of the source code from the itis_taxon
application to provide a model that reflects the taxonomic hierarchy
inherent in taxon codes. This commit includes a model, views and
templates to render lists of taxon objects, an api endpoint including
filter, serializer and list view.  It does include tests yet, but all
existing tests pass.