Removed order clause from taxon_node_filter, ManUnit short_name
This commit updates the custom taxon node filter to remove the
order_by clause which was causing errors in external packages.
Removing the clause does not effect how it works. this commit also
include a new method on Management Unit Object - "short_name" that
includes the management unit type abbreviation, a colon, and the
management unit label.  Used extensively in the templates to display
management units in a complete, but concise manner.  The detail pages
were also given a default geometry (all lakes bounding box) just
in case the lake geometry was not available. all previously passing
tests continue to pass.