Fixed typo in title block, made admin lookup-csv form more specific
This commit fixes a typo in the title block of the main html template,
and updates the admin csv-lookup form to be more specific to common
application - ensure that it is only used to upload lookup values in
the common application and is not used by other applications for other
purposes. All previously passing tests continue to pass.